Use the Number as double when creating a BigDecimal

+++ io.joel/src/main/java/io/joel/impl/TypeConverter.java
@@ -170,7 +170,7 @@ public class TypeConverter {
             return new BigDecimal(bigInteger);
         }
         if (value instanceof Number number) {
-            return new BigDecimal(number.longValue());
+            return new BigDecimal(number.doubleValue());
         }
         if (value instanceof String newValue) {
             return new BigDecimal(newValue);
